competitors_scraper/hitachi/requirements.txt
# HITACHIスクレイパーの依存ライブラリ
selenium>=4.10.0
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
tqdm>=4.66.1
pandas>=2.0.0
webdriver-manager>=4.0.0
